Move sketch files to an appropriately named folder
The Arduino IDE requires that sketches be located in a folder that matches the sketch filename. When you open a sketch file that does not meet this requirement the Arduino IDE asks you whether you'd like the file to be moved to an appropriately named folder. The problem with this is it only moves the file you opened but this project has multiple sketch files. This causes compilation of the sketch to fail:

fatal error: moveservos.h: No such file or directory
 #include "moveservos.h"